Make mode aware of AI Bots, give them a cool name
Check if a player who just joined is an AI Bot by attempting to change their name. If the name change went through, then set their `IsAIBot` to true and `IsPlayer` to false.
The name used contains a unicode character that cannot be used by human players, so this shouldn't have any false positives.

While we are at it, we can set a funny name so they are more easily identified as AI Bots (even though Overwatch's "Ana n" is good enough already)

I plan to use IsAIBot in a subsequent commit.